fix griefer

+++ gamer.hpp
@@ -6,7 +6,7 @@
 namespace Gamer {
 
   void Init();
-  void Update(float time);
+  void Update(long double time);
   void Draw(sf::RenderWindow &window);
 }
